+++ database/dataSource.ts
@@ -93,6 +93,7 @@ import { AjoutDureeMoyenneSejourDansLesProfils1762435683856 } from "./migrations
 import { SupprimeProfilSansLabel1765967970441 } from "./migrations/1765967970441-SupprimeProfilSansLabel";
 import { AjoutTauxOccupationGlobal1765980868528 } from "./migrations/1765980868528-AjoutTauxOccupationGlobal"
 import { UpdateVueRecherche1769680336893 } from "./migrations/1769680336893-UpdateVueRecherche";
+import { AjouterFailedAttemptsEtLockUntil1771495505819 } from "./migrations/1771495505819-AjouterFailedAttemptsEtLockUntil";
 import { updateProfileTable1796422585498 } from "./migrations/1796422585498-updateProfileTable";
 import { AddCreatedByToProfileTable1796792910177 } from "./migrations/1796792910177-AddCreatedByToProfileTable";
 import { ModificationValeurProfil1797341938070 } from "./migrations/1797341938070-modificationValeurProfil";
@@ -379,7 +380,8 @@ const datasource = new DataSource({
     SupressionTableVigierhContrat1763372516438,
     AjouteTypePassageProfessions1764753717478,
     AjoutTauxOccupationGlobal1765980868528,
-    UpdateVueRecherche1769680336893
+    UpdateVueRecherche1769680336893,
+    AjouterFailedAttemptsEtLockUntil1771495505819
   ],
   type: "postgres",
   url: environmentVariables.DATABASE_URL,